fixing assumed profile issus for HELMS3_REGION
When using helm s3 plugin with assumed roles on instances, the switch for just the region does not work, as the existing session is bound to region of the role already. 
This has no impact on the .aws/profile use case as it would not even use the ENV vars which are central to the assume role setups most people have. 

This should fix https://github.com/hypnoglow/helm-s3/issues/128 as well